refactor: 메모 수정, 색
