+++ components/hospital/icu/main/chart/selected-chart/chart-body/chart-body.tsx
@@ -1,7 +1,8 @@
 import ChartInfos from '@/components/hospital/icu/main/chart/selected-chart/chart-body/chart-infos/chart-infos'
-import ChartMemos from '@/components/hospital/icu/main/chart/selected-chart/chart-body/chart-memos/chart-memos'
+import ChartMemos, {
+  type Memo,
+} from '@/components/hospital/icu/main/chart/selected-chart/chart-body/chart-memos/chart-memos'
 import ChartTable from '@/components/hospital/icu/main/chart/selected-chart/chart-body/table/chart-table'
-import { Memo } from '@/hooks/use-memo-management'
 import type { SelectedChart } from '@/types/icu/chart'
 
 export default function ChartBody({ chartData }: { chartData: SelectedChart }) {
