+++ components/hospital/icu/main/chart/selected-chart/chart-body/chart-memos/memo-color-picker.tsx
@@ -6,9 +6,9 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from '@/components/ui/popover'
-import { MEMO_COLORS } from '@/hooks/use-memo-management'
 import { Check } from 'lucide-react'
 import { Dispatch, SetStateAction, useState } from 'react'
+import { MEMO_COLORS } from './chart-memos'
 
 export default function MemoColorPicker({
   memoColor,
