+++ components/hospital/icu/main/chart/selected-chart/chart-body/chart-memos/single-memo/single-memo.tsx
@@ -1,8 +1,8 @@
 import { Textarea } from '@/components/ui/textarea'
 import { toast } from '@/components/ui/use-toast'
-import { MEMO_COLORS, type Memo } from '@/hooks/use-memo-management'
 import { Check, Pencil } from 'lucide-react'
 import React, { useEffect, useRef, useState } from 'react'
+import { type Memo, MEMO_COLORS } from '../chart-memos'
 import MemoColorPicker from '../memo-color-picker'
 import DeleteMemoDialog from './delete-memo-dialog'
 import MemoTimeStamp from './memo-time-stamp'
